Checkpoint: The chunking algorithm (almost) works ! Still a bit brittle (i.e. doesn't get past the first chunk). Things to try - 1. fix the reporting bug in the LBM stack (and which the developers have already shown how to do). 2. Disable all transmits during the FUOTA session. Getting closer...

+++ include/zephyr/smp_lbm/smp_lbm.h
@@ -5,6 +5,6 @@
 #ifndef LBM_ZEPHYR_SMP_LBM_H
 #define LBM_ZEPHYR_SMP_LBM_H
 
-void smp_lbm_downlink(uint8_t port, uint8_t len, const uint8_t *hex_data);
+void smp_lbm_downlink(uint8_t port, uint16_t len, const uint8_t *hex_data);
 
 #endif //LBM_ZEPHYR_SMP_LBM_H
